classical channel updates

+++ programs/2_QuantumKey/AlignmentGUI/runReceiver.py
@@ -149,7 +149,7 @@ class MyWindowClass(QMainWindow, form_class):
 			self.buttonStart.setText("Stop")
 
 			#Initialising the motors
-			self.curr_angle = int(self.motor.get_angle())
+			self.curr_angle = float(self.motor.get_angle())
 			self.offset  = int(self.motor.get_offset())
 			self.update_offset(self.offset)
 			self.update_angle(self.curr_angle)
@@ -222,7 +222,7 @@ class MyWindowClass(QMainWindow, form_class):
 		#Change the absolute angle in GUI and here
 		self.statusbar.showMessage("Updating Angle... Please Wait")
 		self.curr_angle = angle_value
-		self.angleInput.setValue(angle_value)
+		self.angleInput.setValue(angle_value) # QDoubleSpinBox - See guiRcv.ui file
 		self.motor.set_angle(angle_value)
 		self.statusbar.showMessage("Updating Angle... Done")
 		return None
